requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
torch>=1.7.0
ultralytics>=0.0.41
streamlit>=1.22.0
pandas>=1.3.0
matplotlib>=3.5.0
Pillow>=8.0.0
google-cloud-storage>=2.0.0
tensorflow>=2.10.0
altair>=4.0.0 